backend/requirements.txt
asgiref>=3.6.0
Django>=4.2.1
django-cors-headers>=4.3.1
djangorestframework>=3.14.0
djangorestframework-simplejwt==5.3.1
pillow==10.3.0
psycopg2-binary>=2.9.9
PyJWT==2.8.0
pytz>=2024.1
sqlparse>=0.4.4
typing_extensions>=4.10.0



